Move state machine code into a folder, add white spaces and comments.
