+++ src/plan_runner/CMakeLists.txt
@@ -1,13 +1,6 @@
 add_subdirectory(plans)
-find_package(Threads REQUIRED)
+add_subdirectory(state_machine)
 
-add_library(plan_manager_state_machine
-        plan_manager_state_machine.h plan_manager_state_machine.cc
-        state_init.h state_init.cc
-        state_idle.h state_idle.cc
-        state_running.h state_running.cc
-        state_error.h state_error.cc)
-target_link_libraries(plan_manager_state_machine robot_plans)
 
 add_library(iiwa_plan_manager iiwa_plan_manager.h iiwa_plan_manager.cc)
 target_link_libraries(iiwa_plan_manager
