+++ src/plan_runner/state_machine/state_error.h
@@ -4,22 +4,25 @@
 class StateError : public PlanManagerStateBase {
 public:
   static PlanManagerStateBase *Instance();
+
   const PlanBase *GetCurrentPlan(PlanManagerStateMachine *state_machine,
                                  const TimePoint &t_now) const override {
     return nullptr;
   };
+
   void QueueNewPlan(PlanManagerStateMachine *state_machine,
                     std::unique_ptr<PlanBase> plan) override;
+
   void PrintCurrentState(
       const PlanManagerStateMachine *state_machine) const override;
-  [[nodiscard]] PlanExecutionStatus get_plan_execution_status() const override {
-    return PlanExecutionStatus::kError;
-  };
+
   [[nodiscard]] PlanManagerStateTypes get_state_type() const override {
     return PlanManagerStateTypes::kStateError;
   };
+
   void receive_new_status_msg(
       PlanManagerStateMachine *state_machine) const override{};
+
 private:
   StateError() : PlanManagerStateBase("ERROR"){};
   static PlanManagerStateBase *instance_;
